T_COMMENTS_PUB_NOVEL/requirements.txt
adblockparser>=0.7
adslproxy>=2.0.0
aiohttp>=3.3.1
alabaster>=0.7.10
appdirs>=1.4.3
asn1crypto>=0.23.0
async-timeout>=3.0.0
attrs>=17.3.0
Automat>=0.6.0
Babel>=2.5.3
backcall>=0.1.0
base64-util>=0.1
bcrypt>=3.1.4
beautifulsoup4>=4.6.0
bleach>=2.1.3
bs4>=0.0.1
certifi>=2017.7.27.1
cffi>=1.11.2
chainmap>=1.0.2
chardet>=3.0.4
click>=6.7
cloudpickle>=0.5.3
colorama>=0.3.9
combomethod>=1.0.10
constantly>=15.1.0
coverage>=4.4.2
crawler>=0.0.2
cryptography>=2.1.3
cssselect>=1.0.1
cycler>=0.10.0
dask>=0.18.1
dask-xgboost>=0.1.5
decorator>=4.3.0
demjson>=2.2.4
distributed>=1.22.0
Django>=1.11.7
docutils>=0.14
EasyProcess>=0.2.3
entrypoints>=0.2.3
et-xmlfile>=1.0.1
fake-useragent>=0.1.8
Flask>=1.0.2
Flask-MySQLdb>=0.2.0
funcparserlib>=0.3.6
HeapDict>=1.0.0
html5lib>=1.0.1
html5lib-truncation>=0.1.0
hyperlink>=17.3.1
idna>=2.6
idna-ssl>=1.0.1
imagesize>=1.0.0
incremental>=17.5.0
ipykernel>=4.8.2
ipyparallel>=6.1.1
ipython>=6.3.1
ipython-genutils>=0.2.0
ipywidgets>=7.2.1
itsdangerous>=0.24
jdcal>=1.4
jedi>=0.12.0
jieba>=0.39
Jinja2>=2.10
jsonpath>=0.75
jsonschema>=2.6.0
jupyter>=1.0.0
jupyter-client>=5.2.3
jupyter-console>=5.2.0
jupyter-core>=4.4.0
kiwisolver>=1.0.1
lxml>=4.1.1
lxml-xpath2-functions>=0.0.4
MarkupSafe>=1.0
matplotlib>=2.2.2
mistune>=0.8.3
msgpack>=0.5.6
multidict>=4.3.1
mysqlclient>=1.3.12
nbconvert>=5.3.1
nbformat>=4.4.0
nose>=1.3.7
notebook>=5.4.1
nulltype>=2.2.11
numpy>=1.14.2
oauthlib>=2.1.0
olefile>=0.44
openpyxl>=2.5.3
options>=1.4.8
packaging>=17.1
pandas>=0.22.0
pandocfilters>=1.4.2
parameterized>=0.6.1
paramiko>=2.4.1
parsel>=1.2.0
parso>=0.2.0
pickleshare>=0.7.4
Pillow>=4.3.0
pip>=9.0.1
prompt-toolkit>=1.0.15
psutil>=5.4.6
pyasn1>=0.3.7
pyasn1-modules>=0.1.5
pybase64>=0.2.1
pycparser>=2.18
PyDispatcher>=2.0.5
PyExecJS>=1.5.1
Pygments>=2.2.0
pymongo>=3.6.1
PyMySQL>=0.7.11
PyMysqlDB>=0.0.2
PyNaCl>=1.2.1
pyOpenSSL>=17.3.0
pyparsing>=2.2.0
pyquery>=1.4.0
PySocks>=1.6.8
pytesseract>=0.2.0
python-dateutil>=2.6.1
pytz>=2017.3
PyVirtualDisplay>=0.2.1
PyYAML>=3.12
pyzmq>=17.0.0
qiniu>=7.2.2
qt5reactor>=0.5
qtconsole>=4.3.1
queuelib>=1.4.2
quoter>=1.6.8
redis>=2.10.6
redispy>=3.0.0
regex>=2018.2.21
requests>=2.18.4
requests-oauthlib>=1.0.0
scikit-learn>=0.19.1
scipy>=1.0.1
Scrapy>=1.4.0
scrapy-redis>=0.6.8
scrapy-splash>=0.7.2
scrapyd>=1.2.0
scrapyjs>=0.2
selenium>=3.7.0
Send2Trash>=1.5.0
service-identity>=17.0.0
setuptools>=36.6.0
simplegeneric>=0.8.1
simplejson>=3.15.0
six>=1.11.0
snowballstemmer>=1.2.1
sortedcontainers>=2.0.4
Sphinx>=1.7.4
sphinxcontrib-websupport>=1.0.1
splash>=3.2
SQLAlchemy>=1.2.6
stem>=1.6.0
svn>=0.3.45
tblib>=1.3.2
terminado>=0.8.1
testpath>=0.3.1
toolz>=0.9.0
tornado>=5.0.2
traitlets>=4.3.2
tweepy>=3.6.0
twisted>=16.2.0
twitter>=1.18.0
urllib3>=1.22
utils>=0.9.0
w3lib>=1.19.0
wcwidth>=0.1.7
webencodings>=0.5.1
Werkzeug>=0.14.1
wheel>=0.30.0
widgetsnbextension>=3.2.1
win-unicode-console>=0.5
wordcloud>=1.4.1
wrapt>=1.10.11
xlrd>=1.1.0
xlwt>=1.3.0
xpath-expressions>=0.2.2
xpath-generator>=0.2
xpathparser>=1.0.0
xvfbwrapper>=0.2.9
yarl>=1.2.5
zc.buildout>=2.11.2
zict>=0.1.3
zope.interface>=4.5.0
